+++ packages/base/src/features/story/components/SpectaDesktopView.tsx
@@ -3,15 +3,13 @@ import {
   IJupyterGISModel,
   IStorySegmentLayer,
 } from '@jupytergis/schema';
-import React, {
-  RefObject,
-  useEffect,
-  useImperativeHandle,
-  useRef,
-} from 'react';
+import React, { RefObject, useImperativeHandle } from 'react';
 
-import { IStoryViewerPanelHandle } from '@/src/features/story/StoryViewerPanel';
+import StoryViewerPanel, {
+  IStoryViewerPanelHandle,
+} from '@/src/features/story/StoryViewerPanel';
 import { SpectaSegmentListPanel } from '@/src/features/story/components/SpectaSegmentListPanel';
+import { useStoryScrollState } from '@/src/features/story/hooks/useStoryScrollState';
 import SpectaPresentationProgressBar from '@/src/workspace/statusbar/SpectaPresentationProgressBar';
 
 interface ISpectaDesktopViewProps {
@@ -49,35 +47,13 @@ export function SpectaDesktopView({
   showGradient,
   setIndex,
 }: ISpectaDesktopViewProps): JSX.Element {
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-  const topSentinelRef = useRef<HTMLDivElement>(null);
-  const bottomSentinelRef = useRef<HTMLDivElement>(null);
-  const atTopRef = useRef(false);
-  const atBottomRef = useRef(false);
-
-  useEffect(() => {
-    const root = scrollContainerRef.current;
-    const topEl = topSentinelRef.current;
-    const bottomEl = bottomSentinelRef.current;
-    if (!root || !topEl || !bottomEl) {
-      return;
-    }
-    const observer = new IntersectionObserver(
-      (entries: IntersectionObserverEntry[]) => {
-        for (const entry of entries) {
-          if (entry.target === topEl) {
-            atTopRef.current = entry.isIntersecting;
-          } else if (entry.target === bottomEl) {
-            atBottomRef.current = entry.isIntersecting;
-          }
-        }
-      },
-      { root, threshold: 0, rootMargin: '0px' },
-    );
-    observer.observe(topEl);
-    observer.observe(bottomEl);
-    return () => observer.disconnect();
-  }, [currentIndex]);
+  const {
+    scrollContainerRef,
+    topSentinelRef,
+    bottomSentinelRef,
+    getAtTop,
+    getAtBottom,
+  } = useStoryScrollState({ currentIndex });
 
   useImperativeHandle(
     storyViewerPanelRef,
@@ -87,11 +63,11 @@ export function SpectaDesktopView({
       spectaMode: isSpecta,
       hasPrev,
       hasNext,
-      getAtTop: () => atTopRef.current,
-      getAtBottom: () => atBottomRef.current,
+      getAtTop,
+      getAtBottom,
       getScrollContainer: () => scrollContainerRef.current,
     }),
-    [handlePrev, handleNext, isSpecta, hasPrev, hasNext],
+    [handlePrev, handleNext, isSpecta, hasPrev, hasNext, getAtTop, getAtBottom],
   );
 
   return (
@@ -113,7 +89,7 @@ export function SpectaDesktopView({
               data-story-scroll-sentinel="top"
               style={{ height: 1, minHeight: 1, pointerEvents: 'none' }}
             />
-            <SpectaSegmentListPanel
+            {/* <SpectaSegmentListPanel
               model={model}
               isSpecta={isSpecta}
               storyData={storyData}
@@ -123,8 +99,8 @@ export function SpectaDesktopView({
               hasPrev={hasPrev}
               hasNext={hasNext}
               setIndex={setIndex}
-            />
-            {/* <StoryViewerPanel
+            /> */}
+            <StoryViewerPanel
               model={model}
               isSpecta={isSpecta}
               segmentContainerRef={segmentContainerRef}
@@ -137,7 +113,7 @@ export function SpectaDesktopView({
               hasPrev={hasPrev}
               hasNext={hasNext}
               setIndex={setIndex}
-            /> */}
+            />
             <div
               ref={bottomSentinelRef}
               aria-hidden
